# Cloud json marshaling optimization