--- a/stats/cloud/bench_test.go
+++ b/stats/cloud/bench_test.go
@@ -21,6 +21,7 @@
 package cloud
 
 import (
+	"encoding/json"
 	"fmt"
 	"net/url"
 	"testing"
@@ -55,7 +56,7 @@ func BenchmarkAggregateHTTP(b *testing.B) {
 	require.NoError(b, err)
 	now := time.Now()
 	collector.referenceID = "something"
-	var containersCount = time.Duration(500000)
+	containersCount := 500000
 
 	for _, tagCount := range []int{1, 5, 35, 315, 3645} {
 		tagCount := tagCount
@@ -63,34 +64,17 @@ func BenchmarkAggregateHTTP(b *testing.B) {
 			b.ResetTimer()
 			for s := 0; s < b.N; s++ {
 				b.StopTimer()
-				var container = make([]stats.SampleContainer, containersCount)
-				for i := time.Duration(1); i <= containersCount; i++ {
-					var status = "200"
-					if int(i)%tagCount%7 == 6 {
+				container := make([]stats.SampleContainer, containersCount)
+				for i := 1; i <= containersCount; i++ {
+					status := "200"
+					if i%tagCount%7 == 6 {
 						status = "404"
-					} else if int(i)%tagCount%7 == 5 {
+					} else if i%tagCount%7 == 5 {
 						status = "500"
 					}
 
-					container[i-1] = &httpext.Trail{
-						Blocked:        i % 200 * 100 * time.Millisecond,
-						Connecting:     i % 200 * 200 * time.Millisecond,
-						TLSHandshaking: i % 200 * 300 * time.Millisecond,
-						Sending:        i % 200 * 400 * time.Millisecond,
-						Waiting:        500 * time.Millisecond,
-						Receiving:      600 * time.Millisecond,
-						EndTime:        now.Add(i % 100 * 100),
-						ConnDuration:   500 * time.Millisecond,
-						Duration:       i % 150 * 1500 * time.Millisecond,
-						Tags: stats.IntoSampleTags(&map[string]string{
-							"test": "mest", "a": "b",
-							"custom": fmt.Sprintf("group%d", int(i)%tagCount%9),
-							"group":  fmt.Sprintf("group%d", int(i)%tagCount%5),
-							"status": status,
-							"url":    fmt.Sprintf("something%d", int(i)%tagCount%11),
-							"name":   fmt.Sprintf("else%d", int(i)%tagCount%11),
-						}),
-					}
+					tags := generateTags(i, tagCount, map[string]string{"status": status})
+					container[i-1] = generateHTTPExtTrail(now, time.Duration(i), tags)
 				}
 				collector.Collect(container)
 				b.StartTimer()
@@ -98,5 +82,95 @@ func BenchmarkAggregateHTTP(b *testing.B) {
 				collector.bufferSamples = nil
 			}
 		})
+	}
+}
+
+func generateTags(i, tagCount int, additionals ...map[string]string) *stats.SampleTags {
+	res := map[string]string{
+		"test": "mest", "a": "b",
+		"custom": fmt.Sprintf("group%d", i%tagCount%9),
+		"group":  fmt.Sprintf("group%d", i%tagCount%5),
+		"url":    fmt.Sprintf("something%d", i%tagCount%11),
+		"name":   fmt.Sprintf("else%d", i%tagCount%11),
+	}
+	for _, a := range additionals {
+		for k, v := range a {
+			res[k] = v
+		}
+	}
+	return stats.IntoSampleTags(&res)
+}
+
+func BenchmarkMetricMarshal(b *testing.B) {
+	for _, count := range []int{10000, 100000, 500000} {
+		count := count
+		b.Run(fmt.Sprintf("%d", count), func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				b.StopTimer()
+				samples := generateSamples(count)
+				b.StartTimer()
+				r, err := json.Marshal(samples)
+				require.NoError(b, err)
+				b.SetBytes(int64(len(r)))
+			}
+		})
+	}
+}
+
+func generateSamples(count int) []*Sample {
+	samples := make([]*Sample, count)
+	now := time.Now()
+	for i := range samples {
+		tags := generateTags(i, 200)
+		switch i % 3 {
+		case 0:
+			samples[i] = &Sample{
+				Type:   DataTypeSingle,
+				Metric: "something",
+				Data: &SampleDataSingle{
+					Time:  Timestamp(now),
+					Type:  stats.Counter,
+					Tags:  tags,
+					Value: float64(i),
+				},
+			}
+		case 1:
+			aggrData := &SampleDataAggregatedHTTPReqs{
+				Time: Timestamp(now),
+				Type: "aggregated_trend",
+				Tags: tags,
+			}
+			trail := generateHTTPExtTrail(now, time.Duration(i), tags)
+			aggrData.Add(trail)
+			aggrData.Add(trail)
+			aggrData.Add(trail)
+			aggrData.Add(trail)
+			aggrData.Add(trail)
+			aggrData.CalcAverages()
+
+			samples[i] = &Sample{
+				Type:   DataTypeAggregatedHTTPReqs,
+				Metric: "something",
+				Data:   aggrData,
+			}
+		default:
+			samples[i] = NewSampleFromTrail(generateHTTPExtTrail(now, time.Duration(i), tags))
+		}
+	}
+	return samples
+}
+
+func generateHTTPExtTrail(now time.Time, i time.Duration, tags *stats.SampleTags) *httpext.Trail {
+	return &httpext.Trail{
+		Blocked:        i % 200 * 100 * time.Millisecond,
+		Connecting:     i % 200 * 200 * time.Millisecond,
+		TLSHandshaking: i % 200 * 300 * time.Millisecond,
+		Sending:        i % 200 * 400 * time.Millisecond,
+		Waiting:        500 * time.Millisecond,
+		Receiving:      600 * time.Millisecond,
+		EndTime:        now.Add(i % 100 * 100),
+		ConnDuration:   500 * time.Millisecond,
+		Duration:       i % 150 * 1500 * time.Millisecond,
+		Tags:           tags,
 	}
 }