Final increase 
```
name                    old time/op    new time/op     delta
MetricMarshal/10000-8     93.7ms ± 3%     22.6ms ± 7%   -75.83%  (p=0.000 n=8+10)
MetricMarshal/100000-8     893ms ± 3%      210ms ± 9%   -76.52%  (p=0.000 n=10+10)
MetricMarshal/500000-8     4.69s ± 5%      1.41s ±30%   -69.88%  (p=0.000 n=8+10)

name                    old speed      new speed       delta
MetricMarshal/10000-8   44.1MB/s ± 3%  182.8MB/s ± 8%  +314.47%  (p=0.000 n=8+10)
MetricMarshal/100000-8  46.3MB/s ± 3%  197.8MB/s ± 8%  +326.85%  (p=0.000 n=10+10)
MetricMarshal/500000-8  43.5MB/s ±13%  151.1MB/s ±30%  +247.26%  (p=0.000 n=9+10)

name                    old alloc/op   new alloc/op    delta
MetricMarshal/10000-8     28.8MB ± 6%      5.2MB ± 6%   -82.08%  (p=0.000 n=10+10)
MetricMarshal/100000-8     279MB ± 0%       56MB ± 3%   -79.80%  (p=0.000 n=10+9)
MetricMarshal/500000-8    1.29GB ± 0%     0.41GB ± 0%   -67.94%  (p=0.000 n=10+10)

name                    old allocs/op  new allocs/op   delta
MetricMarshal/10000-8       290k ± 0%         0k ± 5%   -99.93%  (p=0.000 n=10+10)
MetricMarshal/100000-8     2.90M ± 0%      0.00M ± 3%   -99.94%  (p=0.000 n=10+9)
MetricMarshal/500000-8     14.5M ± 0%       0.0M ± 0%   -99.91%  (p=0.000 n=10+10)
```
so around 3-4x increase :tada: