@MStoykov 2020-07-24
I highly recommend going commit by commit :D