@MStoykov 2020-07-24
Given the results in 9d30c07 I am wondering if moving to using the marshaling to writer is worth it ... especially because we do copy the body at least once more due to the retry logic ...