@MStoykov 2020-07-27
Both seem to be supported only for unmarshalling ... which is not interesting for our case :D. I am more interested in copying the code and writing it a little bit more readable and just not regenerating it ... but I fear that we will change something and suddenly I will need to rewrite it more - so I am not doing it at least for now :D 

A thing that should benefit from that is the `Sample.Data` is a known implementation always ... but the easyjson code needs to check if it implements an interface ... I doubt this will be a significant improvement though :D