@MStoykov 2020-08-17
@imiric 5d68221 here some results and another benchmark (I might need to prune the number of them later on).